a minor fix to TextAttribute - seems TextAttribute is a read-only PropertyViewAttribute instead a read-write PropertyViewAttribute.
