+++ robobinding/src/main/java/org/robobinding/viewattribute/textview/TextAttribute.java
@@ -15,53 +15,19 @@
  */
 package org.robobinding.viewattribute.textview;
 
-import org.robobinding.property.ValueModel;
-import org.robobinding.viewattribute.AbstractPropertyViewAttribute;
 import org.robobinding.viewattribute.AbstractReadOnlyPropertyViewAttribute;
 
 import android.widget.TextView;
 
 /**
- * 
+ *
  * @since 1.0
  * @version $Revision: 1.0 $
  * @author Robert Taylor
  */
-public class TextAttribute extends AbstractTextAttribute<TextView> {
-
+public class TextAttribute extends AbstractReadOnlyPropertyViewAttribute<TextView, CharSequence> {
     @Override
-    protected AbstractPropertyViewAttribute<TextView, ?> createNewCharSequenceAttribute() {
-	CharSequenceTextAttribute charSequenceTextAttribute = new CharSequenceTextAttribute();
-	return charSequenceTextAttribute;
-    }
-
-    @Override
-    protected AbstractPropertyViewAttribute<TextView, ?> createNewStringAttribute() {
-	StringTextAttribute stringTextAttribute = new StringTextAttribute();
-	return stringTextAttribute;
-    }
-
-    private abstract static class AbstractCharSequenceTextAttribute<PropertyType extends CharSequence> extends
-	    AbstractReadOnlyPropertyViewAttribute<TextView, PropertyType> {
-	@Override
-	protected void valueModelUpdated(PropertyType newValue) {
-	    view.setText(newValue);
-	}
-
-	protected abstract void updateValueModel(ValueModel<PropertyType> valueModel, CharSequence charSequence);
-    }
-
-    static class StringTextAttribute extends AbstractCharSequenceTextAttribute<String> {
-	@Override
-	protected void updateValueModel(ValueModel<String> valueModel, CharSequence charSequence) {
-	    valueModel.setValue(charSequence.toString());
-	}
-    }
-
-    static class CharSequenceTextAttribute extends AbstractCharSequenceTextAttribute<CharSequence> {
-	@Override
-	protected void updateValueModel(ValueModel<CharSequence> valueModel, CharSequence charSequence) {
-	    valueModel.setValue(charSequence);
-	}
+    protected void valueModelUpdated(CharSequence newValue) {
+	view.setText(newValue);
     }
 }
